Resolve 3 (of many) clang compiler warnings
Reviewers: #popart, fabiant

Reviewed By: #popart, fabiant

Subscribers: shauryas, vidhyasudhanl

Maniphest Tasks: T19855

Differential Revision: https://phabricator.sourcevertex.net/D24993

+++ willow/include/popart/iarray.hpp
@@ -9,11 +9,11 @@ namespace popart {
 class IArray {
 public:
   virtual ~IArray() {}
-  virtual void *data()                 = 0;
-  virtual DataType dataType() const    = 0;
-  virtual std::size_t rank() const     = 0;
-  virtual int64_t dim(int index) const = 0;
-  virtual std::size_t nelms() const    = 0;
+  virtual void *data()                    = 0;
+  virtual DataType dataType() const       = 0;
+  virtual std::size_t rank() const        = 0;
+  virtual int64_t dim(size_t index) const = 0;
+  virtual std::size_t nelms() const       = 0;
 };
 
 } // namespace popart
